Update the package name to reflect the moved package after refactoring
